Change yaml serialization to use scalar and support class detection
Switch serialization to a simple scalar with the standard python
content.  This requires that we can reconstruct the Config
subclass from it. We now do this by parsing the first two
lines of the string to determine the relevant class to import.
This is a class method such that other serialization schemes
can reuse it.